refactor(preorders): split god service into domain sub-services
Extract PreordersService (822 lines) into three focused sub-services under
services/ — PreordersCrudService (CRUD + queries), PreordersStatusService
(state machine + loyalty points side-effects), PreordersFinancialService
(receipt generation). PreordersService becomes a 75-line orchestrator that
delegates to sub-services while preserving the exact same public API, all
invariants (state machine, ledger writes, multi-tenant scoping), and module
exports unchanged.
